refactor #4 - extract out detail xml view
